fix: resolve Rust 1.95 clippy lints, mutex poisoning, and library-parity host check
- builddiag-fix: drop redundant `.into_iter()` (clippy::useless_conversion).
- builddiag-cli/bdd: drop `as usize` no-op casts (clippy::unnecessary_cast).
- builddiag-app: recover poisoned git test mutex (parallel-run safety) and
  disable GPG signing in test git commands so CI without GPG passes.
- xtask: normalize host.os and host.arch in the conformance comparison so the
  library-parity check works regardless of which OS regenerated the goldens.

All tests pass, clippy clean with `-D warnings`, and `xtask conform` shows
10/10 passing.

+++ crates/builddiag-cli/tests/bdd/steps.rs
@@ -517,15 +517,15 @@ fn then_sensor_report_exists_at_canonical_path(world: &mut BuilddiagWorld) {
     }
 
     assert_eq!(
-        report.verdict.counts.info as usize, info,
+        report.verdict.counts.info, info,
         "sensor verdict info count should match findings"
     );
     assert_eq!(
-        report.verdict.counts.warn as usize, warn,
+        report.verdict.counts.warn, warn,
         "sensor verdict warn count should match findings"
     );
     assert_eq!(
-        report.verdict.counts.error as usize, error,
+        report.verdict.counts.error, error,
         "sensor verdict error count should match findings"
     );
 }
